backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.6
cymem==2.0.11
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
fsspec==2025.7.0
gunicorn==23.0.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.3.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
preshed==3.0.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.0
python-decouple==3.8
python-docx==1.2.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.17.2
